requirements.txt
chardet>=5.0.0
openpyxl>=3.0.0
fpdf2>=2.7.4
pandas>=2.0.0
streamlit>=1.22.0